refactor(ui): extract coordinator-admin/data/actions.ts — 5 async action handlers via factory
Extract createGroupFromAdminPanel, runGroupAction, createInviteFromAdminPanel,
reviewJoinRequestFromAdminPanel, runDeviceAction into a factory module.

createCoordinatorAdminActions({ renderShell, reloadData }) returns the 5 handlers.
The barrel tab wires in lazy arrow closures so the hoisted renderShell and
loadCoordinatorAdminData references resolve at call time.
